Rename IdentityRegistration to IdentityMenu to unify naming
