+++ src/renderer/assets/less/4_components/identity-menu.less
@@ -69,11 +69,11 @@
   }
 }
 
-#identity-registration {
+#identity-menu {
   left: auto;
   right: 0;
 
-  .identity-registration-content {
+  .identity-menu-content {
     padding: 0 1.1em 1.1em 1.1em;
     border-bottom: none;
   }
