+++ src/renderer/assets/less/index.less
@@ -58,7 +58,7 @@
 @import '4_components/stats';
 @import '4_components/tutorial';
 @import '4_components/feedback-form';
-@import '4_components/identity-registration';
+@import '4_components/identity-menu';
 @import '4_components/icon-ethereum.less';
 @import '4_components/icon-logo.less';
 @import '4_components/connection-history';
